Revert "Merge branch 'Minestom:master' into ticks-and-update-velocity-access-modifiers"
This reverts commit 130397a1fca94fed0b10d16beaf8c83ef1ef360e, reversing
changes made to 64d55a9944ccd37c0dfa22576f59a43eaa663759.

+++ src/main/java/net/minestom/server/entity/metadata/display/TextDisplayMeta.java
@@ -43,8 +43,8 @@ public class TextDisplayMeta extends AbstractDisplayMeta {
         super.metadata.setIndex(OFFSET + 2, Metadata.VarInt(value));
     }
 
-    public byte getTextOpacity() {
-        return super.metadata.getIndex(OFFSET + 3, (byte) -1);
+    public int getTextOpacity() {
+        return super.metadata.getIndex(OFFSET + 3, -1);
     }
 
     public void setTextOpacity(byte value) {
